Exponential Moving Average
With the help of my professor, I have figured out how to use the EMA function with my data to find the exponential moving average of the sales of each of the items. I plan to use this code to perform similar calculations with different forms of moving averages over different time frames and then compare them to find which one is best. I may also use different algorithms depending on the volatility of sales.